Withdraw resize for test.
